requirements.txt
astroid==2.5
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
isort==5.7.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pyglet==1.5.0
pylint==2.7.1
scipy==1.6.1
toml==0.10.2
typed-ast==1.4.2
Werkzeug==1.0.1
wrapt==1.12.1

